fix: Use a slug to identify models
When we update our models we're pushing them up to a file named after
the version we're deploying.  When we update our models the endpoint
configuration doesn't seem to notice the change.  We're assuming that
this is because there's no noticable change to the endpoint
configuration because we're only changing an attribute of the model.

To force such changes to take effect we're changing the name of the
model resource that is create, and we're doing that based on the
filename of the model object stored in S3.